+++ src/Documentation/MetadataTypes.ts
@@ -1,4 +1,4 @@
 /**
  * Type used to define a Class with a constructor.
  */
-export type Constructor = { new (...args: any[]): any };
+export type Constructor<T = any> = { new (...args: any[]): T };